refactor(recruit-board): 패키지 구조 변경(query, command 패키지 제거)
