+++ src/main/java/com/somemore/domains/recruitboard/controller/RecruitBoardCommandApiController.java
@@ -1,36 +1,29 @@
 package com.somemore.domains.recruitboard.controller;
 
 
-import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
-
+import com.somemore.domains.recruitboard.dto.request.RecruitBoardCreateRequestDto;
+import com.somemore.domains.recruitboard.dto.request.RecruitBoardLocationUpdateRequestDto;
+import com.somemore.domains.recruitboard.dto.request.RecruitBoardStatusUpdateRequestDto;
 import com.somemore.domains.recruitboard.dto.request.RecruitBoardUpdateRequestDto;
-import com.somemore.domains.recruitboard.usecase.command.CreateRecruitBoardUseCase;
-import com.somemore.domains.recruitboard.usecase.command.DeleteRecruitBoardUseCase;
-import com.somemore.domains.recruitboard.usecase.command.UpdateRecruitBoardUseCase;
+import com.somemore.domains.recruitboard.usecase.CreateRecruitBoardUseCase;
+import com.somemore.domains.recruitboard.usecase.DeleteRecruitBoardUseCase;
+import com.somemore.domains.recruitboard.usecase.UpdateRecruitBoardUseCase;
 import com.somemore.global.auth.annotation.CurrentUser;
 import com.somemore.global.common.response.ApiResponse;
 import com.somemore.global.imageupload.dto.ImageUploadRequestDto;
 import com.somemore.global.imageupload.usecase.ImageUploadUseCase;
-import com.somemore.domains.recruitboard.dto.request.RecruitBoardCreateRequestDto;
-import com.somemore.domains.recruitboard.dto.request.RecruitBoardLocationUpdateRequestDto;
-import com.somemore.domains.recruitboard.dto.request.RecruitBoardStatusUpdateRequestDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import java.time.LocalDateTime;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
 
 @Tag(name = "Recruit Board Command API", description = "봉사 활동 모집글 생성 수정 삭제 API")
 @RequiredArgsConstructor
