+++ src/main/java/com/somemore/domains/recruitboard/usecase/CreateRecruitBoardUseCase.java
@@ -1,6 +1,7 @@
-package com.somemore.domains.recruitboard.usecase.command;
+package com.somemore.domains.recruitboard.usecase;
 
 import com.somemore.domains.recruitboard.dto.request.RecruitBoardCreateRequestDto;
+
 import java.util.UUID;
 
 public interface CreateRecruitBoardUseCase {
