+++ src/main/java/com/somemore/domains/recruitboard/usecase/DeleteRecruitBoardUseCase.java
@@ -1,4 +1,4 @@
-package com.somemore.domains.recruitboard.usecase.command;
+package com.somemore.domains.recruitboard.usecase;
 
 import java.util.UUID;
