+++ src/test/java/com/somemore/domains/recruitboard/service/UpdateRecruitBoardServiceTest.java
@@ -1,4 +1,4 @@
-package com.somemore.domains.recruitboard.service.command;
+package com.somemore.domains.recruitboard.service;
 
 import com.somemore.domains.location.domain.Location;
 import com.somemore.domains.location.repository.LocationRepository;
